Store the dispatch_list in ETS, not application:set_env
Use a protected, named table, so everyone can read it, but only
webmachine_router can write to it. The owner of the table is
webmachine_router with the heir of the process being the supervisor.

If webmachine_router is restarted, the supervisor gives ownership of the
table to the newly spawned webmachine_router process again. This means
that routing configuration is not lost if webmachine_router crashes.